Import CirrOS with Dockerfile!
Changing the mkimage script to a fetcher,
just so you know where we're getting this image from.

Because ADD requires a "local" tar.gz file, we must
keep a copy of the downloaded image in this git repository.
That's pretty lame, but there isn't another alternative besides
using wrapping shell scripts (and it's better to just have
a git repo you can checkout and run 'docker build' against!)